% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear_fit}
\alias{linear_fit}
\title{linear}
\usage{
linear_fit(x, y)
}
\arguments{
\item{x}{input independent data matrix - nxp}

\item{y}{input dependent vector - nx1}
}
\value{
estimated beta, fitted value, residual
}
\description{
This function estimates linear regression model with OLS.
The output is supposed to be exactly same as lm().
}
\examples{

library(penaltyRegression)
set.seed(0)
x <- matrix(runif(50), ncol = 1)
y <- matrix(2*x[,1] + rnorm(50,mean = 0, sd=0.1), ncol = 1)

res_linear = linear_fit(x,y)
res_linear$beta
res_linear$fit
res_linear$residual
reg_plot(x, y, res_linear$beta)
res_plot(res_linear$residual, res_linear$fit)

lm(y~x-1)$coef
}
