% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reg_plot}
\alias{reg_plot}
\title{reg_plot}
\usage{
reg_plot(x, y, beta)
}
\arguments{
\item{x}{input data frame for independent variable - nx1}

\item{y}{input array for dependent variable - n}

\item{beta}{input coefficients - p}
}
\value{
scatter plot
}
\description{
This function aims to create y-x scatter plot.
If input col > 1, it will only make y~x1 plot
}
\examples{
library(penaltyRegression)
set.seed(0)
x <- matrix(runif(50), ncol = 1)
y <- matrix(2*x[,1] + rnorm(50,mean = 0, sd=0.1), ncol = 1)
res_linear = linear_fit(x,y) #fit linear regression
reg_plot(x, y, res_linear$beta) #y-x scatter plot
}
