% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reg_predict}
\alias{reg_predict}
\title{reg_predict}
\usage{
reg_predict(x, beta)
}
\arguments{
\item{x}{input independent data matrix - nxp}

\item{beta}{input coefficients - p}
}
\value{
estimated beta with L1 regularization
}
\description{
This function aims to predict linear model, OLS / L2 / L1 regularization
}
\examples{


}
