% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{res_plot}
\alias{res_plot}
\title{res_plot}
\usage{
res_plot(residual, fitted)
}
\arguments{
\item{residual}{input data frame for independent variable - n}

\item{fitted}{input array for dependent variable - n}
}
\value{
scatter plot
}
\description{
This function aims to create residual-fitted scatter plot.
}
\examples{
set.seed(0)
library(penaltyRegression)
x <- matrix(runif(50), ncol = 1)
y <- matrix(2*x[,1] + rnorm(50,mean = 0, sd=0.1), ncol = 1)
res_linear = linear_fit(x,y) #fit linear regression
res_plot(res_linear$residual, res_linear$fit) #residual plot
}
