% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reg_plot}
\alias{reg_plot}
\title{reg_plot}
\usage{
reg_plot(x, y, beta)
}
\arguments{
\item{x}{input data frame for independent variable - nx1}

\item{y}{input array for dependent variable - n}

\item{beta}{input coefficients - p}
}
\value{
scatter plot
}
\description{
This function aims to create y-x scatter plot.
If input col > 1, it will only make y~x1 plot
}
\examples{

}
