% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{res_plot}
\alias{res_plot}
\title{res_plot}
\usage{
res_plot(residual, fitted)
}
\arguments{
\item{residual}{input data frame for independent variable - n}

\item{fitted}{input array for dependent variable - n}
}
\value{
scatter plot
}
\description{
This function aims to create residual-fitted scatter plot.
}
\examples{

}
